TEP-1015: refactor price conf, report settings
Push price related configuration to Settings.
Refactor report settings to kernel-level module, next to the settings.
