+++ tackler-core/src/kernel.rs
@@ -3,19 +3,24 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-///
-/// Predicate to test if item x is part of set or not
-///
-/// `x` item to be tested
-/// `returns` true if it's selected, false if it's rejected
 pub use settings::Settings;
 pub(crate) mod accumulator;
 pub mod balance;
 pub mod hash;
 pub mod price_lookup;
 pub mod report_item_selector;
+pub mod report_settings;
 pub mod settings;
 
+pub use report_settings::BalanceGroupSettings;
+pub use report_settings::BalanceSettings;
+pub use report_settings::RegisterSettings;
+
+///
+/// Predicate to test if item x is part of set or not
+///
+/// `x` item to be tested
+/// `returns` true if it's selected, false if it's rejected
 pub trait Predicate<T> {
     fn eval(&self, item: &T) -> bool;
 }
