+++ tackler-core/src/report/balance_group_reporter.rs
@@ -3,51 +3,18 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use crate::kernel::accumulator;
 use crate::kernel::accumulator::TxnGroupByOp;
-use crate::kernel::price_lookup::PriceLookup;
 use crate::kernel::report_item_selector::BalanceSelector;
-use crate::kernel::Settings;
-use crate::model::{Commodity, Transaction, TxnSet};
+use crate::kernel::{accumulator, BalanceGroupSettings};
+use crate::kernel::{BalanceSettings, Settings};
+use crate::model::{Transaction, TxnSet};
+use crate::report::BalanceReporter;
 use crate::report::{write_acc_sel_checksum, write_report_timezone, Report};
-use crate::report::{BalanceReporter, BalanceSettings};
-use crate::{config::Scale, model::price_entry::PriceDb};
 use jiff::tz::TimeZone;
+use std::error::Error;
 use std::io;
-use std::{error::Error, sync::Arc};
 use tackler_api::txn_ts;
 use tackler_api::txn_ts::GroupBy;
-
-#[derive(Debug, Clone)]
-pub struct BalanceGroupSettings {
-    pub title: String,
-    pub ras: Vec<String>,
-    pub group_by: GroupBy,
-    pub report_tz: TimeZone,
-    pub report_commodity: Option<Arc<Commodity>>,
-    pub price_lookup: PriceLookup,
-    pub scale: Scale,
-}
-
-impl BalanceGroupSettings {
-    pub fn from(
-        settings: &Settings,
-        group_by: Option<GroupBy>,
-        report_commodity: Option<Arc<Commodity>>,
-        price_lookup: PriceLookup,
-    ) -> Result<BalanceGroupSettings, Box<dyn Error>> {
-        let bgs = BalanceGroupSettings {
-            title: settings.report.balance_group.title.clone(),
-            ras: settings.get_balance_group_ras(),
-            group_by: group_by.unwrap_or(settings.report.balance_group.group_by),
-            report_tz: settings.report.report_tz.clone(),
-            report_commodity,
-            price_lookup,
-            scale: settings.report.scale.clone(),
-        };
-        Ok(bgs)
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct BalanceGroupReporter {
@@ -87,14 +54,13 @@ impl Report for BalanceGroupReporter {
         cfg: &Settings,
         writer: &mut W,
         txn_data: &TxnSet<'_>,
-        price_db: &PriceDb,
     ) -> Result<(), Box<dyn Error>> {
         let bal_acc_sel = self.get_acc_selector()?;
 
         let price_lookup_ctx = &self.report_settings.price_lookup.make_ctx(
             &txn_data.txns,
             self.report_settings.report_commodity.clone(),
-            price_db,
+            &cfg.price.price_db,
         );
 
         let group_by_op = self.get_group_by_op();
@@ -121,7 +87,7 @@ impl Report for BalanceGroupReporter {
             title: String::default(),
             ras: vec![],
             scale: self.report_settings.scale.clone(),
-            commodity: self.report_settings.report_commodity.clone(),
+            report_commodity: self.report_settings.report_commodity.clone(),
             price_lookup: self.report_settings.price_lookup.clone(),
         };
         for bal in &bal_groups {
